+++ modules/objects.js
@@ -1,6 +1,7 @@
+/* eslint-disable linebreak-style */
 export default class Book {
-    constructor(title, author) {
-      this.title = title;
-      this.author = author;
-    }
+  constructor(title, author) {
+    this.title = title;
+    this.author = author;
   }
+}
