+++ modules/variables.js
@@ -1,0 +1,20 @@
+/* eslint-disable linebreak-style */
+/* eslint-disable no-trailing-spaces */
+/* eslint-disable indent */
+/* eslint-disable linebreak-style */
+/* eslint-disable no-multiple-empty-lines */
+
+const listBook = document.getElementById('book-list');
+const addForm = document.getElementById('add-form');
+const bookListing = document.getElementById('listBook');
+const bookReg = document.getElementById('addBook');
+const contact = document.getElementById('contact-section');
+
+const listLink = document.getElementById('list-book');
+const addLink = document.getElementById('new-book');
+const contactLink = document.getElementById('contact');
+const date = document.getElementById('date');
+
+export {
+ listBook, addForm, bookListing, bookReg, contact, listLink, addLink, contactLink, date, 
+};
